requirement.txt
Django==4.0.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.1.0
django-filter==2.4.0
drf-yasg==1.20.0